Revert "Сделала работу с лайками через сервер."
This reverts commit 2bdf33ce6783313930b965a14359998752bdd8e8.

+++ android-app/app/src/main/java/ru/netology/nmedia/service/DateTimeService.kt
@@ -8,7 +8,7 @@ import java.util.Locale
 object DateTimeService {
     // Конвертирует и форматирует UnixTime в строку с датой и временем публикации
     fun formatUnixTime(unixDateTime: Long): String {
-        val instant = Instant.ofEpochSecond(unixDateTime)
+        val instant = Instant.ofEpochSecond(unixDateTime.toLong())
         val formatter = DateTimeFormatter.ofPattern("dd MMMM yyyy 'в' HH:mm", Locale("ru"))
             .withZone(ZoneId.systemDefault())
         return formatter.format(instant)
